Refactor tags
Remake association from one:many to many:many for posts and tags.
Create admin actions create, update, delete for tag. Moved all
controllers to namespace api.
