--- app/controllers/admin/comments_controller.rb
@@ -1,9 +1,0 @@
-module Admin
-  class CommentsController < ApplicationController
-    def destroy
-      comment = Comment.find(params[:id])
-      authorize comment
-      comment.destroy
-    end
-  end
-end
