--- app/controllers/admin/posts_controller.rb
@@ -1,9 +1,0 @@
-module Admin
-  class PostsController < ApplicationController
-    def destroy
-      post = Post.find(params[:id])
-      authorize post
-      post.destroy
-    end
-  end
-end
