--- app/controllers/admin/users_controller.rb
@@ -1,9 +1,0 @@
-module Admin
-  class UsersController < ApplicationController
-    def destroy
-      user = User.find(params[:id])
-      authorize user
-      user.destroy
-    end
-  end
-end
