+++ app/controllers/api/admin/comments_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module Admin
+    class CommentsController < ApplicationController
+      def destroy
+        comment = Comment.find(params[:id])
+        authorize comment
+        comment.destroy
+      end
+    end
+  end
+end
