+++ app/controllers/api/admin/posts_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module Admin
+    class PostsController < ApplicationController
+      def destroy
+        post = Post.find(params[:id])
+        authorize post
+        post.destroy
+      end
+    end
+  end
+end
