+++ app/controllers/api/admin/users_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module Admin
+    class UsersController < ApplicationController
+      def destroy
+        user = User.find(params[:id])
+        authorize user
+        user.destroy
+      end
+    end
+  end
+end
