+++ app/controllers/application_controller.rb
@@ -3,10 +3,15 @@ class ApplicationController < ActionController::API
   include Pagy::Backend
 
   before_action :authenticate_user
+  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
   attr_reader :current_user
 
   protected
+
+  def user_not_authorized
+    render json: { error: 'You are not authorized to perform this action' }
+  end
 
   def authenticate_user
     @current_user = user
