+++ app/models/post.rb
@@ -5,7 +5,8 @@ class Post < ApplicationRecord
 
   belongs_to :user
 
-  has_many :tags, dependent: :destroy
+  has_many :tagging, dependent: :destroy
+  has_many :tags, through: :tagging
   has_many :comments, dependent: :destroy
   has_many :images, as: :imageable, dependent: :destroy
   has_many :rates, as: :rateable, dependent: :destroy
@@ -13,7 +14,7 @@ class Post < ApplicationRecord
 
   scope :fresh, -> { where('posts.created_at >= ?', now - 1.day) }
   scope :hot, -> { left_joins(:comments).where('comments.created_at >= ?', now - 1.day).group(:id) }
-  scope :best, -> { left_joins(:rates).where('rates.created_at >= ?', Post.now - 1.day).where(rates: { status: 'like' }).group(:id) }
+  scope :best, -> { left_joins(:rates).where('rates.created_at >= ?', now - 1.day).where(rates: { status: 'like' }).group(:id) }
   scope :order_by_likes, -> { left_joins(:rates).where(rates: { status: 'like' }).group(:id).order('count(rates.id) desc') }
   scope :order_by_created, -> { order(created_at: :desc) }
 end
