+++ app/models/tagging.rb
@@ -1,0 +1,6 @@
+class Tagging < ApplicationRecord
+  validates :post, uniqueness: { scope: :tag }
+
+  belongs_to :post
+  belongs_to :tag
+end
