+++ app/policies/tag_policy.rb
@@ -1,0 +1,13 @@
+class TagPolicy < ApplicationPolicy
+  def create?
+    user.admin?
+  end
+
+  def update?
+    user.admin?
+  end
+
+  def destroy?
+    user.admin?
+  end
+end
