+++ db/migrate/20200702071212_create_tags.rb
@@ -1,9 +1,7 @@
 class CreateTags < ActiveRecord::Migration[6.0]
   def change
     create_table :tags do |t|
-      t.integer :post_id, null: false
       t.string :tag, null: false
     end
-    add_index :tags, :post_id
   end
 end
