+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_07_03_071656) do
+ActiveRecord::Schema.define(version: 2020_07_06_032130) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -66,10 +66,14 @@ ActiveRecord::Schema.define(version: 2020_07_03_071656) do
     t.index ["user_id", "rateable_id", "rateable_type"], name: "index_rates_on_user_id_and_rateable_id_and_rateable_type", unique: true
   end
 
-  create_table "tags", force: :cascade do |t|
+  create_table "taggings", force: :cascade do |t|
     t.integer "post_id", null: false
+    t.integer "tag_id", null: false
+    t.index ["post_id", "tag_id"], name: "index_taggings_on_post_id_and_tag_id", unique: true
+  end
+
+  create_table "tags", force: :cascade do |t|
     t.string "tag", null: false
-    t.index ["post_id"], name: "index_tags_on_post_id"
   end
 
   create_table "users", force: :cascade do |t|
